Module Beagle_prey

module Words : sig ... end

type words = Words.t

type t

val caught : (t * words) Bap_primus.Std.Primus.observation

val detected : t Bap_primus.Std.Primus.observation

made when a prey is detected

val finished : t Bap_primus.Std.Primus.statement

a statement that a prey was detected

val catch : (t * words) Bap_primus.Std.Primus.statement

val create : Bap.Std.tid Bap.Std.seq -> string -> t

create terms chars the result of beagle hunting stating that during the execution of a sequence of terms we observed chars in that specific order.

val data : t -> string

data prey is a sequence of chars, that was caught

val terms : t -> Bap.Std.tid Bap.Std.seq

terms prey a sequence of terms that, when executed, consumed or produced one of the data prey characters.

Attributes that are added by beagle analysis.

val chars : words Bap.Std.tag

Each string in a set is a sequence of characters that were detected by Beagle on each emulation (it is possible, that beagle will sniff the same term more than once). The characters are specified in an order in which they were observed.

val strings : words Bap.Std.tag

a set of static strings that we directly or indirectly referenced the emulation of a term.

val words : words Bap.Std.tag

a set of words that can be built from a specified alphabet with the observed characters.

val statics : string Bap.Std.Addr.Map.t Bap.Std.tag

statically discovered strings