Module Beagle_prey.Words

type t = Core_kernel.String.Set.t

include Bap.Std.Value.S with type t := t

type t

In order to construct a value with the a given type you must provide an implementation for marshaling functions, comparison function and pretty-printing.

val bin_t : t Bin_prot.Type_class.t

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_shape_t : Bin_prot.Shape.t

val compare : t -> t -> int

val t_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val pp : Stdlib.Format.formatter -> t -> unit

val to_string : t -> string